+++ lib/screen_lock_plugin_method_channel.dart
@@ -5,6 +5,7 @@ import 'screen_lock_plugin_platform_interface.dart';
 class MethodChannelScreenLockPlugin extends ScreenLockPluginPlatform {
   @visibleForTesting
   final methodChannel = const MethodChannel('screen_lock_plugin');
+  final EventChannel _eventChannel = const EventChannel('screen_lock_plugin/events');
 
   @override
   Future<bool?> lockScreen() async {
@@ -23,5 +24,16 @@ class MethodChannelScreenLockPlugin extends ScreenLockPluginPlatform {
   @override
   Future<void> requestDeviceAdmin() async {
     await methodChannel.invokeMethod<void>('requestDeviceAdmin');
+  }
+
+  @override
+  Future<bool?> isScreenOn() async {
+    final result = await methodChannel.invokeMethod<bool>('isScreenOn');
+    return result;
+  }
+
+  @override
+  Stream<String> onScreenStateChanged() {
+    return _eventChannel.receiveBroadcastStream().map((event) => event as String);
   }
 }